refactor: Articles api
